+++ src/ITurnstile.sol
@@ -3,5 +3,5 @@ pragma solidity ^0.8.0;
 
 interface ITurnstile {
     function register(address) external returns (uint256);
-    function assign(uint) external returns (uint);
+    function assign(uint256) external returns (uint256);
 }
